4.15-4.23

+++ part4/bloglist/controller/blogs.js
@@ -1,23 +1,32 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 require('express-async-errors')
 
+
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { blogs: 0 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   if (!request.body.title || !request.body.url)  {
     response.status(400).end()
   } else {
-    const blog = new Blog(request.body)
+    const user = request.user
+    const blog = new Blog(Object.assign(request.body, { user: user.id }))
     const result = await blog.save()
+    user.blogs = user.blogs.concat(result._id)
+    await user.save()
     response.status(201).json(result)
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() !== request.user.id) {
+    return response.status(401).json({ error: 'no permission to delete blog' })
+  }
   const result = await Blog.findByIdAndDelete(request.params.id)
   response.json(result)
 })
